+++ hephaistos-data/lua/FixScreenComponents.lua
@@ -39,6 +39,7 @@ end
 
 Import "../Mods/Hephaistos/Filters/AwardMenuScripts.lua"
 Import "../Mods/Hephaistos/Filters/CodexScripts.lua"
+Import "../Mods/Hephaistos/Filters/GhostAdminScreen.lua"
 Import "../Mods/Hephaistos/Filters/MetaUpgrades.lua"
 Import "../Mods/Hephaistos/Filters/QuestLogScreen.lua"
 Import "../Mods/Hephaistos/Filters/RunHistoryScreen.lua"